requirements.txt
Flask==2.0.1
Pillow==10.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==20.1.0 